Clean up more compact renderer files
