+++ src/renderer/src/components/mobile/mobile-platform-copy.ts
@@ -1,0 +1,17 @@
+import type { Platform } from './MobileHero'
+
+export const PLATFORM_COPY: Record<
+  Platform,
+  { description: string; ctaLabel: string; url: string }
+> = {
+  ios: {
+    description: 'Scan with your iPhone camera to open the App Store.',
+    ctaLabel: 'Open App Store',
+    url: 'https://apps.apple.com/app/orca-ide/id6766130217'
+  },
+  android: {
+    description: 'Scan with your Android camera to download the latest APK from GitHub Releases.',
+    ctaLabel: 'Download APK',
+    url: 'https://github.com/stablyai/orca/releases/download/mobile-v0.0.11/app-release.apk'
+  }
+}
